seperate AboutsPage sections into files
