+++ packages/ui/src/atoms/Divider/Divider.tsx
@@ -1,4 +1,5 @@
-import React, { forwardRef } from "react"
+import type React from "react"
+import { forwardRef } from "react"
 import { cn } from "../../utils"
 
 interface DividerProps extends React.HTMLAttributes<HTMLDivElement> {
